requirements.txt
python-pptx>=1.0.0
pandas>=2.0.0
lxml>=4.9.0
pillow>=9.0.0
numpy>=1.24.0
streamlit>=1.20.0
openpyxl>=3.0.0
xlsxwriter>=3.0.0
requests>=2.28.0
python-docx>=1.1.0
